Update to my updated change
Forgot to add "useDependencyInformation" for the "dependencies" to actually be used.
This along with the last patch to the ChickenMod.java file should fix the ability to use other mod items. It'll have the other mods load before this one so that the items exist to be used. The reason behind this is here https://github.com/setycz/ChickensMod/issues/153